Add serializer
